requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
albumentations==0.5.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blis==0.7.4
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
effdet==0.2.1
efficientnet-pytorch==0.6.3
entrypoints==0.3
fastai==2.1.4
fastcore==1.3.2
fastprogress==1.0.0
fsspec==0.8.5
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
icevision==0.5.2
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==3.1.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206598566/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.5.1
jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148959479/work
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multidict==5.1.0
munch==2.5.0
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nose==1.3.7
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.7
pandas==1.1.5
pandocfilters==1.4.3
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
pretrainedmodels==0.7.4
prometheus-client==0.9.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
psutil==5.8.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.1.5
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
requests-oauthlib==1.3.0
resnest==0.0.6b20210124
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.4
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
sentry-sdk==0.19.5
Shapely==1.7.1
shortuuid==1.0.1
six @ file:///tmp/build/80754af9/six_1605205313296/work
smmap==3.0.5
spacy==2.3.4
srsly==1.0.5
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.3
threadpoolctl==2.1.0
tifffile==2021.1.14
timm==0.3.2
torch==1.7.1
torchvision==0.8.2
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.54.1
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.7.4.3
urllib3==1.26.2
wandb==0.10.15
wasabi==0.8.0
watchdog==0.10.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.4.0
